# docs/openapi.json
{
  "components": {
    "schemas": {
      "CreateSessionRequest": {
        "properties": {
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Limits"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CreateSessionRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "last_activity_at": {
            "title": "Last Activity At",
            "type": "integer"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "CREATING",
              "RUNNING",
              "IDLE",
              "STOPPED",
              "DESTROYING",
              "DESTROYED"
            ],
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "tenant_id",
          "limits",
          "created_at",
          "last_activity_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nControl plane for an LLM-agent sandbox service. See\n[SPECIFICATION.md](https://github.com/JISUlicious/sandboxing/blob/main/SPECIFICATION.md)\nand\n[ARCHITECTURE.md](https://github.com/JISUlicious/sandboxing/blob/main/ARCHITECTURE.md)\nfor the full contract.\n\nEvery session is a long-lived, hardened container. Filesystem state in\n`/workspace` persists across exec calls and across stop/resume; process\nstate does not (SPEC-002). All endpoints under `/v1/sessions` require a\nbearer token in the `Authorization` header.\n",
    "title": "Sandbox Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "description": "Creates a long-lived sandbox session under the calling tenant. Starts in `RUNNING` with an empty `/workspace` (SPEC-101). `limits` are merged with tenant-default values; per-field caps are enforced (SPEC-100, SPEC-300).",
        "operationId": "create_session_v1_sessions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Tenant concurrency or per-field limit exceeded (`limit_exceeded`). SPEC \u00a76."
          }
        },
        "summary": "Create a session",
        "tags": [
          "Sessions"
        ]
      }
    },
    "/v1/sessions/{session_id}": {
      "delete": {
        "description": "Removes the container and its `/workspace` volume. Irreversible (SPEC-105). The destroy is multi-step (DESTROYING \u2192 DESTROYED) and idempotent on partial failure (ARCH-051).",
        "operationId": "destroy_session_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session not found, not owned by the calling tenant, or already DESTROYED (SPEC-200)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Destroy a session",
        "tags": [
          "Sessions"
        ]
      },
      "get": {
        "operationId": "get_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session not found, not owned by the calling tenant, or already DESTROYED (SPEC-200)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a session's status and limits",
        "tags": [
          "Sessions"
        ]
      }
    },
...
    "/v1/sessions/{session_id}/resume": {
      "post": {
        "description": "Restarts the container for a `STOPPED` or `IDLE` session. Process state from before the stop is gone \u2014 only filesystem state persists (ARCH \u00a73.4).",
        "operationId": "resume_session_v1_sessions__session_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session not found, not owned by the calling tenant, or already DESTROYED (SPEC-200)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session is in a state that doesn't allow this operation (`invalid_state`)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume a stopped session",
        "tags": [
          "Sessions"
        ]
      }
    },
...
  },
...
}